fix: fixed example list bug - it was created everytime a user logged in
To fix it, I had to create a parameter in UserRepository, it receives now a SharedPreference var. It permits the example list to be created only once, when the user logs by the first time. This change affected the singleton UserRepository, so I had to create an intermediate file called ServiceLocator to handle with only one instance of the UserRepository in the app